Binomial initial implementation
