+++ frontend/src/components/OptiPrice/Binomial/BinomialChart.jsx
@@ -1,105 +1,155 @@
 import React from "react";
-import BinomialChart from "./BinomialChart";
+import {
+  LineChart,
+  Line,
+  XAxis,
+  YAxis,
+  Tooltip,
+  ResponsiveContainer,
+  CartesianGrid,
+} from "recharts";
 
-function BinomialVisual({
-  convergence = [],
-  precision = "advanced",
-  onPrecisionChange,
-  onGenerate,
-  price,
-  form,
-  loading,
-}) {
-  const precisionOptions = [
-    { label: "Low Precision", value: "simple", desc: "Logarithmic steps" },
-    {
-      label: "Medium Precision",
-      value: "advanced",
-      desc: "Linear + Log steps",
-    },
-    { label: "High Precision", value: "precise", desc: "All steps" },
-  ];
+function CustomTooltip({ active, payload }) {
+  if (active && payload && payload.length) {
+    return (
+      <div className="bg-white/95 backdrop-blur-sm border border-gray-200 p-3 rounded-lg shadow-lg">
+        <div className="text-sm text-gray-600">Steps (N)</div>
+        <div className="font-medium text-gray-900">{payload[0].payload.N}</div>
+        <div className="text-sm text-gray-600 mt-1">Option Price</div>
+        <div className="font-semibold text-blue-600">
+          ${payload[0].payload.price}
+        </div>
+      </div>
+    );
+  }
+  return null;
+}
 
-  const isCall = form?.option_type === "call";
-  const optionTypeText = isCall ? "Call Option" : "Put Option";
-  const isAmerican = form?.style === "american" || form?.american === true;
-  const styleText = isAmerican ? "American" : "European";
+// Simplified smart ticks (powers of 2 + min/max)
+function getSmartTicks(data) {
+  if (!data || !data.length) return [];
+  const minN = data[0].N;
+  const maxN = data[data.length - 1].N;
+  if (data.length <= 20) return data.map((d) => d.N);
+  const ticks = [minN];
+  let pow = 1;
+  while (pow < maxN) {
+    if (pow > minN) ticks.push(pow);
+    pow *= 2;
+  }
+  if (!ticks.includes(maxN)) ticks.push(maxN);
+  return Array.from(new Set(ticks)).sort((a, b) => a - b);
+}
+
+function BinomialChart({ data, isCall }) {
+  if (!data || data.length === 0) {
+    return (
+      <div className="space-y-4">
+        <div className="text-center">
+          <h3 className="text-lg font-semibold text-gray-900 mb-1">
+            Binomial Tree Convergence
+          </h3>
+          <p className="text-sm text-gray-600">
+            How option price converges as tree steps increase
+          </p>
+        </div>
+        <div className="h-80 bg-gray-50 rounded-xl border border-gray-200 flex items-center justify-center">
+          <p className="text-gray-500">
+            Click "Generate Convergence Analysis" to see the chart
+          </p>
+        </div>
+      </div>
+    );
+  }
+
+  const lastN = data[data.length - 1]?.N;
+  const ticks = getSmartTicks(data);
 
   return (
-    <div className="space-y-6">
+    <div className="space-y-4">
       <div className="text-center">
-        <h2 className="text-2xl font-bold text-gray-900 mb-2">
-          Binomial Model
-        </h2>
-        <p className="text-gray-600">
-          Lattice-based approach for {styleText.toLowerCase()} option pricing
-        </p>
-      </div>
-
-      {/* Price Display */}
-      <div className="bg-gradient-to-r from-blue-50 to-indigo-50 rounded-2xl p-6 border border-blue-200">
-        <div className="text-center">
-          <p className="text-sm font-medium text-blue-600 mb-1">
-            {styleText} {optionTypeText} Price
-          </p>
-          <p className="text-3xl font-bold text-blue-900">
-            {typeof price === "number" ? `$${price.toFixed(4)}` : "—"}
-          </p>
-          <p className="text-xs text-blue-600 mt-2">
-            Calculated at N = 512 steps
-          </p>
-        </div>
-      </div>
-
-      {/* Precision Settings */}
-      <div className="bg-white rounded-2xl border border-gray-200 p-6 shadow-sm">
-        <h3 className="text-lg font-semibold text-gray-900 mb-4">
-          Convergence Analysis Settings
+        <h3 className="text-lg font-semibold text-gray-900 mb-1">
+          {isCall ? "Call" : "Put"} Option Price Convergence
         </h3>
-        <p className="text-sm text-gray-600 mb-4">
-          Choose precision level for convergence visualization. Higher precision
-          shows more data points.
+        <p className="text-sm text-gray-600">
+          How option price converges as binomial tree steps increase
         </p>
-
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-3">
-          {precisionOptions.map((option) => (
-            <button
-              key={option.value}
-              onClick={() => onPrecisionChange(option.value)}
-              disabled={loading}
-              className={`p-4 rounded-xl border text-left transition-all duration-200 ${
-                precision === option.value
-                  ? "bg-gradient-to-r from-blue-50 to-indigo-50 border-blue-300 ring-2 ring-blue-500/20"
-                  : "bg-white border-gray-200 hover:border-gray-300 hover:bg-gray-50"
-              } ${loading ? "opacity-50 cursor-not-allowed" : ""}`}
-            >
-              <div className="font-medium text-gray-900">{option.label}</div>
-              <div className="text-sm text-gray-600 mt-1">{option.desc}</div>
-            </button>
-          ))}
-        </div>
-
-        <button
-          onClick={onGenerate}
-          disabled={loading}
-          className={`w-full mt-6 py-3 px-4 rounded-xl font-medium transition-all duration-200 ${
-            loading
-              ? "bg-gray-200 text-gray-400 cursor-not-allowed"
-              : "bg-gradient-to-r from-blue-600 to-indigo-600 text-white hover:from-blue-700 hover:to-indigo-700 shadow-md hover:shadow-lg"
-          }`}
-        >
-          {loading
-            ? "Generating Convergence Data..."
-            : "Generate Convergence Analysis"}
-        </button>
       </div>
 
-      {/* Chart */}
-      <div className="bg-white rounded-2xl border border-gray-200 p-6 shadow-sm">
-        <BinomialChart data={convergence} isCall={isCall} />
+      <div className="h-80 w-full">
+        <ResponsiveContainer width="100%" height="100%">
+          <LineChart
+            data={data}
+            margin={{ top: 20, right: 30, left: 20, bottom: 60 }}
+          >
+            <CartesianGrid
+              strokeDasharray="3 3"
+              stroke="#e5e7eb"
+              opacity={0.5}
+            />
+            <XAxis
+              dataKey="N"
+              label={{
+                value: "Steps (N)",
+                position: "insideBottom",
+                offset: -40,
+                style: { textAnchor: "middle", fill: "#6b7280" },
+              }}
+              stroke="#6b7280"
+              tick={{ fill: "#6b7280", fontSize: 12 }}
+              tickLine={{ stroke: "#6b7280" }}
+              type="number"
+              domain={[data[0].N, data[data.length - 1].N]}
+              ticks={ticks}
+              allowDecimals={false}
+            />
+            <YAxis
+              label={{
+                value: "Option Price ($)",
+                angle: -90,
+                position: "insideLeft",
+                style: { textAnchor: "middle", fill: "#6b7280" },
+              }}
+              stroke="#6b7280"
+              tick={{ fill: "#6b7280", fontSize: 12 }}
+              tickLine={{ stroke: "#6b7280" }}
+              type="number"
+              domain={["auto", "auto"]}
+            />
+            <Tooltip content={<CustomTooltip />} />
+            <Line
+              type="monotone"
+              dataKey="price"
+              stroke={isCall ? "#3b82f6" : "#ef4444"}
+              strokeWidth={3}
+              dot={false}
+              name={`${isCall ? "Call" : "Put"} Price`}
+              strokeLinecap="round"
+            />
+            {/* Highlight the final N as a gold dot */}
+            <Line
+              type="monotone"
+              dataKey="price"
+              stroke="#facc15"
+              strokeWidth={0}
+              dot={({ cx, cy, payload }) =>
+                payload.N === lastN ? (
+                  <circle key={payload.N} cx={cx} cy={cy} r={5} fill="#facc15" />
+                ) : null
+              }
+              legendType="none"
+            />
+          </LineChart>
+        </ResponsiveContainer>
+      </div>
+
+      <div className="text-center">
+        <p className="text-xs text-gray-500">
+          Chart shows binomial tree convergence - gold dot indicates final price at N=512
+        </p>
       </div>
     </div>
   );
 }
 
-export default BinomialVisual;
+export default BinomialChart;
