+++ frontend/src/components/OptiPrice/BlackScholes/BlackScholesVisual.jsx
@@ -14,18 +14,16 @@ function BlackScholesVisual({ form }) {
         </p>
       </div>
 
-      <div className="bg-white rounded-2xl border border-gray-200 p-6 shadow-sm">
-        <BlackScholesFormula
-          S0={form.S0}
-          K={form.K}
-          T={form.T}
-          r={form.r}
-          sigma={form.sigma}
-          q={form.q}
-          includeDividend={form.includeDividend}
-          option_type={form.option_type}
-        />
-      </div>
+      <BlackScholesFormula
+        S0={form.S0}
+        K={form.K}
+        T={form.T}
+        r={form.r}
+        sigma={form.sigma}
+        q={form.q}
+        includeDividend={form.includeDividend}
+        option_type={form.option_type}
+      />
 
       <div className="bg-white rounded-2xl border border-gray-200 p-6 shadow-sm">
         <BlackScholesChart form={form} />
